Extract FakeRefreshControl class to a separate file
