+++ EssentialFeed/EssentialFeediOSTests/FeedViewControllerTests.swift
@@ -459,17 +459,3 @@ extension FeedViewController {
         refreshControl = fake
     }
 }
-private class FakeRefreshcontrol: UIRefreshControl {
-    private var _isRefreshing = false
-    
-    override var isRefreshing: Bool { _isRefreshing }
-    
-    override func beginRefreshing() {
-        _isRefreshing = true
-    }
-    
-    override func endRefreshing() {
-        _isRefreshing = false
-    }
-}
-
